function_app/requirements.txt
azure-functions==1.21.3
azure-functions-durable==1.2.9
azure-identity==1.19.0
azure-mgmt-authorization==4.0.0
azure-monitor-ingestion==1.0.4
azure-core==1.32.0
httpx==0.27.0
